+++ convex/auth.config.ts
@@ -1,14 +1,10 @@
 import type { AuthConfig } from "convex/server";
 
-declare const process: {
-  env: Record<string, string | undefined>;
-};
-
 export default {
   providers: [
     {
       domain: "https://accounts.google.com",
-      applicationID: process.env.GOOGLE_CLIENT_ID!,
+      applicationID: "365609603908-modldahk3205acfdf1pshhckufho13v0.apps.googleusercontent.com",
     },
   ],
 } satisfies AuthConfig;